Spar tests (#539)
* Fix: `stack test --fast --test-arguments=--seed=762030435 services/spar`

* simplify

* Make spar log "listening on ..." on startup.

* Test: SCIM and SAML work together and SCIM-created users can login.

* Refactor: keep less vulnerable state in TestSpar, recreate it in each test instead.

* Fix: some tests were messing with the test environment, confusing other tests.

* Fix: create new SAML user for SCIM posts, not just brig.

* Fix: run forgotten spar integration tests.

* Fix: make forgotten spar integration tests actually work.

* Cleanup: names.

* haddocks.

* Make SCIM test token construction random; un-sort imports.
