+++ services/spar/package.yaml
@@ -125,3 +125,4 @@ executables:
       - xml-conduit
       - xml-hamlet
       - xml-lens
+      - zauth
