+++ services/spar/test-integration/Spec.hs
@@ -12,9 +12,11 @@ import Test.Hspec
 import Util
 
 import qualified Data.Yaml as Yaml
-import qualified Test.Spar.DataSpec
-import qualified Test.Spar.SCIMSpec
 import qualified Test.Spar.APISpec
+import qualified Test.Spar.AppSpec
+import qualified Test.Spar.DataSpec
+import qualified Test.Spar.Intra.BrigSpec
+import qualified Test.Spar.SCIMSpec
 
 
 main :: IO ()
@@ -28,10 +30,11 @@ mkspec = do
   serviceOpts :: Opts <- Yaml.decodeFileEither configFilePath >>= either (throwIO . ErrorCall . show) deriveOpts
 
   pure . beforeAll (mkEnv integrationOpts serviceOpts) . afterAll destroyEnv $ do
-    describe "Test.Spar.Data" Test.Spar.DataSpec.spec
-    describe "Test.Spar.SCIM" Test.Spar.SCIMSpec.spec
-    -- This should be the last test, because it deletes the test IdP.
     describe "Test.Spar.API" Test.Spar.APISpec.spec
+    describe "Test.Spar.App" Test.Spar.AppSpec.spec
+    describe "Test.Spar.Data" Test.Spar.DataSpec.spec
+    describe "Test.Spar.Intra.Brig" Test.Spar.Intra.BrigSpec.spec
+    describe "Test.Spar.SCIM" Test.Spar.SCIMSpec.spec
 
 
 -- | Accept config file locations as cli options.
