+++ services/spar/test-integration/Test/Spar/DataSpec.hs
@@ -21,7 +21,8 @@ import SAML2.WebSSO as SAML
 import Spar.Data as Data
 import Spar.Types
 import URI.ByteString.QQ (uri)
-import Util
+import Util.Core
+import Util.Types
 
 
 spec :: SpecWith TestEnv
@@ -29,7 +30,7 @@ spec = do
   describe "TTL" $ do
     it "works in seconds" $ do
       env <- ask
-      let idpid = env ^. teIdP . SAML.idpId
+      (_, _, (^. SAML.idpId) -> idpid) <- registerTestIdP
       (_, req) <- call $ callAuthnReq (env ^. teSpar) idpid
 
       let probe :: (MonadIO m, MonadReader TestEnv m) => m Bool
